refactor: controller, domain, view 디렉토리를 생성해서 각 클래스에 해당하는 디렉토리로 클래스를 옮김
