+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.RacingGameController;
+
 public class Application {
 	public static void main(String[] args) {
 		// TODO 구현 진행
